Split deserializer and deserializer_core with some temp glue between them
